Tidy up project with regards to maintainability
PSR-12 is out now, we should use this as a standard and have a
`phpcs` configuration to be able to easily run and hint to the
IDE on what does and does not constitute a formatting error.

Correct the folder names to align with PSR-4 autoloading standard

Ensure the requirements encompass all necessary functionality

Make an attempt to clarify the documentation accompanying the module.