Add tests to ensure functionality is correct as expected
During preparation for adding an additional application, a business
team discovered that the content being added for this new system was
not being read correctly from the API. This is due to page names
not being unique across applications, and content designed for a
page in the new application was being associated with a page of the
same name in an existing application - thus did not show with
requests for content by the new application.

To ensure this issue is remedied and does not regress, tests have
been added to assert that functionality is correct, and that no
regressions outside the area of concern  have occurred while
updating the functionality.